api/app/requirements.txt
astroid==1.6.4
atomicwrites==1.1.5
attrs==18.1.0
backports.functools-lru-cache==1.5
click==6.7
configparser==3.5.0
enum34==1.1.6
Flask==1.0.2
Flask-Login==0.4.1
funcsigs==1.0.2
futures==3.2.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.2.0
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.3
PyJWT==1.6.4
pylint==1.9.1
pytest==3.6.0
singledispatch==3.4.0.3
six==1.11.0
validate-email==1.3
Werkzeug==0.14.1
wrapt==1.10.11
